main_expCOMET only contains the processing of Part 8 (P8), which is 'FiberZwartDik_lang_open'. The others will be added when I can either loop or function what needs to happen. Data is averaged, plotted and then fitted. Coefficients, RMSE, R2 and ErrorRange are then retrieved.
For fitting the data an 8 term multiexponential fit was used, as this worked best when looking at RMSE, R2 and ErrorRange. 

SLM toolbox and Optimization toolbox need to be downloaded to run the file. 

In ALA_data_prep not much has changed, names of the parameters have changed as I was trying to move them into a loop or function (I now know this is not needed).